change: clean up duplication
