+++ src/script.js
@@ -2,4 +2,6 @@ import './styles.scss';
 import './scripts/blockMods/imageAttr';
 import './scripts/registerPlugin';
 import './scripts/helpers/validationHooks';
-import './scripts/blockChecks/coreBlockValidation';
+import './scripts/blockChecks/buttonValidation';
+import './scripts/blockChecks/imageValidation';
+import './scripts/blockChecks/tableValidation';
